refactor: address rnetser review comments
- Move workers_utility_pods + control_plane_utility_pods from
  cluster/nodes.py to cluster/utilities.py (alongside utility_daemonset)
- Move ocp_current_version + openshift_current_version from
  cluster/infrastructure.py back to tests/conftest.py (temporary;
  will be homed in operator/)
